requirements.txt
lightning
numpy
prometheus-fastapi-instrumentator
locust
